win32: don't set error mode
Commit a8c63f25b3 (win32: improve filesystem detection and display)
added a call to SetErrorMode(SEM_FAILCRITICALERRORS).  This was
on the strength of the documentation for GetVolumeInformation()
which suggests that otherwise a message box will appear to prompt
the user to put media in an empty floppy or CD drive.  This would
disrupt the expected behaviour of applets like 'df'.

In practice it seems the call to SetErrorMode() is unnecessary.
It also results in other errors going unreported.

Remove the call to SetErrorMode().

Saves 8-20 bytes.

(GitHub issue #423)