ash: allow HISTFILE=/dev/null to work as intended
It was noted that setting HISTFILE=/dev/null in upstream BusyBox
prevented shell history from being saved to a history file.  This
failed in busybox-w32 with an error from lseek(2).

The lseek(2) wrapper was rather conservative in the file types it
accepted.  Allowing FILE_TYPE_CHAR makes it possible to seek on
/dev/null, thus avoiding the issue with HISTFILE.

In addition, the wrapper for open(2) now converts the Unix-style
mode argument to Windows-style.

(GitHub issue #425)