win32: properly restore BB_ env vars
Some BB_ shell variables get special treatment:  they're updated in
the environment immediately on any change.  One case was missed:
if such a variable was unset or not exported and was overridden by
a local variable it wasn't unset in the environment when the local
variable went out of scope.

Add the code required to do this.

Adds 48-64 bytes.

(GitHub issue #423)