ash: restore value of imported variable on unexport
Shell variables imported from the environment are marked with a
special flag and backslashes in their values are (by default)
replaced with forward slashes.  Sometimes this may not be what
we want.

Modify the 'export' shell built-in so unexporting a variable of
this type restores its original value from the environment and
removes its special flag.  It can then be re-exported.

Adds 32 bytes.

(GitHub issue #428)