make: allow pragmas to apply recursively
Pragmas set when pdpmake is run are exported to the environment
variable PDPMAKE_PRAGMAS as a space-separated list of pragma names.
This environment variable is read when pdpmake starts and any
pragmas it contains are applied.

Thus pragmas are passed to recursive invocations of pdpmake.

PDPMAKE_PRAGMAS can also be set by the user.

Adds 240-288 bytes.