ash: ignore hidden/iconified state
Some third-party terminal programs have a hidden console host to
interact with console applications.  When the busybox-w32 shell
was used with such a terminal the hidden console host could be
revealed because the shell expected it to be iconified.

Since it doesn't much matter in this case whether the console host
is hidden or iconified treat these states as equivalent.

The user's preference set by the 'noiconify' option will still be
respected when a console window is concealed.

(GitHub issue #430)