ash: fix slow running when background job is present
It was noted that a simple 'while' loop would take considerably
longer to run in a shell which had spawned a background job
compared to one that hadn't.

The problem originates with commit 4d8a277d59 (Implement nonblocking
wait) which introduced a nonblocking wait in ash.

This was found to cause the shell to use 100% of CPU when running
'sleep 60'.  This was 'fixed' by commit 88b8cb61e (ash: Add a very
small timeout to nonblocking wait).

Subsequently commit 96c9c0044 (ash: allow waitpid_child to block)
fixed a problem with the logic of the original commit, but it left
the small timeout in place.  It is this timeout which slows down
the shell when a background job is present.

The solution is to restore the 0 timeout for the nonblocking wait.

Saves 0-16 bytes.

(GitHub issue #434)