dd: add support for writing to physical drives
A common use of 'dd' is to copy boot images to removable media.
This didn't work on Windows because opening a physical drive
failed with the default flags used by 'dd'.

If the output file is a physical drive remove incompatible flags.

Writing to a physical drive requires elevated privileges and the
drive must be offline.

Adds 80 bytes.

(GitHub issue #435)