su: add test mode
Testing some uses of 'su' can be challenging because any errors
appear in the new console window which may close as a result.

Add the '-t' option to enable test mode.  This starts a new shell
using ShellExecuteEx(), but without elevated privileges and a new
console.  All other options and arguments are handled much as
before, though some differences in behaviour are to be expected due
to the lack of elevated privilege.

Adds 80-96 bytes.

(GitHub issue #438)