make: disallow inference rules for phony targets
GNU make doesn't allow inference rules or the .DEFAULT target to
be used for phony targets.

POSIX is unclear on the matter but there doesn't seem to be an
explicit prohibition of inference rules or .DEFAULT.

Follow the GNU make behaviour as a non-POSIX extension.

Adds 48-80 bytes.

(pdpmake GitHub issue 56)