ash: allow additional element in argv array
tryexec() in ash relies on all callers of shellexec() having an
additional unused element before the start of its argv array.

In busybox-w32 this was not the case when shellexec() was called
from forkshell_shellexec(), as only the actual contents of the
argv array were copied into the forkshell data block.  In practice
argv[-1] is only currently used when the 'which' applet is about
to be run, so whatever got overwritten (probably cmdtable[30])
was unlikely to matter.

Still, let's be correct and allocate the additional element.

Adds 16 bytes.