which,ash: code shrink detection of standalone shell
Commit 6d87be4d7 (which,ash: changes to which/command/type) let
'which' detect when it's run from a standalone shell by having
the shell pass the undocumented '-s' option.

A better solution is to alter argv[0] when 'which' is run from a
standalone shell.  This is possible because the code path through
run_noexec_applet_and_exit() and which_main() doesn't actually
use argv[0].

- No special treatment is required in ash when 'which' has no
  arguments or the '--help' option.

- tryexec() no longer needs an extra element before the start of
  argv.  The commit 027fb22e2 can be reverted and the allocation
  of argv in evalcommand() simplified.

- 'which' no longer needs to handle the '-s' option.

Saves 96-104 bytes.