ash: optimise running of scripts
The BusyBox shell detects certain cases where forking a command is
unnecessary (last command in a script or subshell, for example) and
calls execve(2) instead.  This doesn't help in the Windows port
because execve(2) is implemented by creating a process.

There is one case where it is possible to apply this optimisation:
if the command is a script and the script interpreter is an applet.

- Have evalcommand() pass a flag to indicate this situation to
  shellexec().  Also, allocate two spare elements before the start
  of the argv array.

- If the flag is TRUE shellexec() passes the shell's PATH variable
  down to tryexec() so it can perform a test for applet override.

- If tryexec() finds that all the necessary conditions apply it
  can run a script by directly invoking the interpreter's main().

Adds 192-224 bytes.