make: duplicate makefile name recorded with command
Commit f3f72ac1d (make: show location of errors during build)
stored a pointer to the name of the makefile with each command,
for use in diagnostic messages.

While this is fine for makefiles defined on the command line, the
names of included files will have been freed before they can be
used.

Always take a copy of the makefile name stored with commands.

Adds 32-48 bytes.

(GitHub issue #449)