libbb: code shrink supplementary group test
Recent upstream changes to file permission tests added a function
to check and cache values in the supplementary group list.

The implementation of getgroups() in the Windows port adds no
useful information beyond what can be obtained by checking the
current effective gid, which all callers of the new function
already do.  The function can be replaced with a simple 'FALSE'.

Saves 232-288 bytes.