win32: fix regression in chdir(2)
When mingw_chdir() was introduced it canonicalised its argument
(585d17d26).  This had the side effect of making its case match
that of the directory as stored on disk.  Subsequent changes
retained that behaviour for symlinks but not otherwise (69d328022,
b99032390).  This was noted to affect the appearance of the
directory specified by the (undocumented) 'sh -d' option.

Fix the case of non-symlink directories too.

Adds 16-32 bytes.