win32: fix problem interrupting shell loop
It proved to be almost impossible to interrupt a loop like:

   while true; do sleep 1; done

where 'sleep' was an external program, not an applet.  The issue
was introduced by commit 0475b7a64 (win32: convert exit codes).
This passed a POSIX error code to the exit() in wait_for_child()
so a parent was unable to detect when its child was interrupted.

Pass the Windows exit code to exit() instead.  Work around the
changes introduced by commit 790e377273 (win32: revert 'don't set
error mode').

Adds 16-32 bytes.