win32: workaround for pipe writability in poll(2)
A CGI script was found to hang when a large amount of data was
posted:

   #!/bin/sh
   echo "Content-type: text/plain;"
   echo
   if [ "$REQUEST_METHOD" = "POST" ]; then
       dd of=my.dat bs=1 count=${CONTENT_LENGTH}
       echo -n "success."
   else
       echo -n "error!"
   fi

This appears to be due to problems determining whether a pipe is
writable on Windows.  The Git for Windows project has a workaround
in their copy of GNUlib's poll(2) implementation.  The details are
in the commit message:

   https://github.com/git-for-windows/git/commit/94f4d01932279c419844aa708bec31a26056bc6b

Apply the same workaround here.

Saves 220-272 bytes.

(GitHub issue #468)