ash: fix waiting for status of background job
The following has never worked properly in busybox-w32:

   $ my_func() { return 5; }
   $ my_func & sleep 1; wait $!; echo $?

The expected result is that 'echo' should display '5'.  Actual
results used to be '0' and more recently have been '127'.

The culprit was commit fa6f44ea72 (win32: ash: reimplement
waitpid(-1)).  When the status of a job changed its pid was set to
-1, which flagged processes of interest to the implementation of
waitpid().

This is no longer necessary as waitpid() now uses the process
handle instead.  There's therefore no need to overwrite the pid.

(GitHub issue #470)