tar: try harder to detect old tar files
The code to autodetect compressed tar files failed to detect a
bunzip2-compressed archive.  When tar was invoked with the 'j'
option it worked fine.

The autodetection code looks for the magic string 'ustar' or a
series of five NULs to determine that an archive is uncompressed.
The failing archives had more than five NULs in the header and
were taken to be uncompressed.

Look for a longer run of NULs: 16 is certainly sufficient for the
archives in question.

Adds 8-16 bytes.

(GitHub issue #475)