make: support GNU/BSD suffixes and inference rules
The specification of inference rules in POSIX implies that only
suffixes starting with a period and containing no other periods
are to be considered.

In contrast, both GNU and BSD make allow suffixes to contain an
arbitrary number of periods, or none at all.  Allow this as an
extension.

Adds 640-816 bytes.

(pdpmake GitHub issue 70)