make: phony targets and double colon rules
Commit 770bad1fe (make: disallow inference rules for phony targets)
added the GNU/BSD make feature that inference rules aren't used
for phony targets.

Normally a double-colon rule without associated commands causes
an inference rule search.  If the target of the double-colon rule
was phony no search was carried out and the rule failed.  As a
result of this perl failed to build (on Linux, I didn't try on
Windows).

When a double-colon rule has no commands and a phony target the
prerequisites should be built and the rule should succeed.

Adds 32 bytes.

(pdpmake GitHub issue #70)