win32: enhanced support for SIGPIPE
Commit 68ddd4ec3c (win32: emulate SIGPIPE) allowed broken pipes
to be detected so that processes could either return EPIPE or
terminate.

Later, commit 1b2ee3667 (win32: add fake HUP and QUIT signals)
added support for two additional fake signals, thus allowing the
'kill' applet to refer to them by name and the shell to define
(though not execute) traps for them.

Add PIPE to the list of signals known to 'kill' and the shell,
though still without support for calling the trap function.

(GitHub issue #482)