win32: Do not open hidden files for executability
If you execute `ls` in the home directory and you have Cortex XDR installed,
the program is aborted and quarantined. that's because busybox-w32
tries to open all files in the directory, even the hidden ones, and Cortex
creates some random files that are hidden.

This commit extends the `starts with #!` rule to `Visible files that start with #!`
when determining if a file is executable.