Further tweaks to make_single_applets.sh
Changes to the applet configurations of 'reset' and 'which' in the
Microsoft Windows port resulted in these applets being enabled in
every single-applet executable built by make_single_applets.sh.

Alter the configuration to avoid this.

Remove PLATFORM_POSIX and PLATFORM_MINGW32 from the list of applets
to be built.

(GitHub issue #536)