mkdir: correctly report directory creation
It was found that 'mkdir -p --verbose C:/' incorrectly reported
that 'C:/' was created.

Commit 0e0b5e9ff8 (mingw.c: ignore EACCES from mkdir if directory
exists) fixed a failure to create directories but caused the
`mkdir` applet to issue the wrong diagnostic.

The underlying problem is in the Windows runtime implementation
of _mkdir(): it can return the wrong errno in some cases, EACCES
instead of EEXIST.

Having the mkdir(2) wrapper return the correct errno should fix
both the directory creation and verbose diagnostic problems.

The website referenced in the previous commit is no longer
available.  archive.org has a copy:

   https://web.archive.org/web/20150224075238/http://www.apijunkie.com/APIJunkie/blog/post/2009/12/22/_mkdir-C-runtime-library-function-might-return-unexpected-error-values.aspx

(GitHub issue #546)