ash: allow ctrl-c to interrupt read in mintty
The command 'read -t 10' couldn't be interrupted by ctrl-c
when ash was running in the mintty terminal emulator.

This issue was introduced by commits 8e6991733 and b2901ce8e
which fixed other problems with the 'read' builtin.

Rearrange the code to avoid calling poll(2) in an interactive
shell on Windows.

(GitHub issue #547)