win32: update UTF-8 manifest
It seems that the UTF-8 manifest can be adjusted to avoid binaries
failing to run on Windows XP.

Make the necessary adjustment.  This doesn't affect the legacy
builds, only those with Unicode enabled.  Nor does it change the
code which prevents Unicode builds from running on platforms
that don't support the manifest.

See this GitHub issue for details:

   https://github.com/skeeto/w64devkit/issues/315